When running tests there is no need for them and they slow down specs.

Non scientific benchmark, running `rspec spec/features/` in hitobito_sac_cas:

Before:
```
Finished in 3 minutes 41.8 seconds (files took 5.14 seconds to load)
140 examples
```
After
```
Finished in 3 minutes 5.7 seconds (files took 3.95 seconds to load)
140 examples
```